Bump PHP version >= 5.5, use PSR-4, change namespace
