+++ .travis.yml
@@ -4,8 +4,6 @@ language: php
 sudo: false
 
 php:
-  - "5.3"
-  - "5.4"
   - "5.5"
   - "5.6"
   - '7.0'
