--- phpspec.yml
@@ -1,4 +1,0 @@
-suites:
-  UniqueIndicator:
-    namespace: Crowdin
-    src_path: lib
